pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "concierge"
version = "0.1.0"
description = "Personal Automation Concierge - GitHub activity monitoring with rule-based notifications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "PowerSchill" }
]
keywords = ["github", "automation", "notifications", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-httpx>=0.30.0",
    "pytest-cov>=4.1.0",
    "freezegun>=1.2.0",
    "ruff>=0.1.0",
    "pyright>=1.1.350",
]

[project.scripts]
concierge = "concierge.cli:app"

[project.urls]
Homepage = "https://github.com/PowerSchill/automation_concierge"
Repository = "https://github.com/PowerSchill/automation_concierge"

[tool.hatch.build.targets.wheel]
packages = ["src/concierge"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

# Ruff configuration (linting and formatting)
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value comparison
]

[tool.ruff.lint.isort]
known-first-party = ["concierge"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Pyright configuration (type checking)
[tool.pyright]
pythonVersion = "3.11"
pythonPlatform = "All"
typeCheckingMode = "strict"
include = ["src"]
exclude = ["tests", "**/__pycache__"]
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true
reportPrivateUsage = "warning"
# Relax some strict rules for generic type inference
reportUnknownVariableType = "warning"
reportUnknownMemberType = "warning"
reportUnknownArgumentType = "warning"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/concierge"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true
